requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
google-genai==1.73.1
python-dotenv==1.1.0
